# pool/frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Convos Pool API",
    "version": "1.0.0",
    "description": "Pool manager API for provisioning, managing, and monitoring Convos agent instances."
  },
  "servers": [{ "url": "/" }],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "POOL_API_KEY as Bearer token"
      },
      "GatewayToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-gateway-token",
        "description": "Per-instance gateway token (set via x-gateway-token header or instanceId + gatewayToken in body)"
      }
    },
...
  },
  "paths": {
...
    "/api/proxy/info": {
      "get": {
        "tags": ["Proxy"],
        "summary": "Get instance's provisioned resources (email, phone)",
        "security": [{ "GatewayToken": [] }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "instanceId": { "type": "string" }, "email": { "type": "string", "nullable": true }, "phone": { "type": "string", "nullable": true } } } } } } }
      }
    },
    "/api/proxy/email/provision": {
      "post": {
        "tags": ["Proxy"],
        "summary": "Provision an AgentMail inbox (idempotent)",
        "security": [{ "GatewayToken": [] }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "email": { "type": "string" }, "provisioned": { "type": "boolean" } } } } } } }
      }
    },
    "/api/proxy/sms/provision": {
      "post": {
        "tags": ["Proxy"],
        "summary": "Provision a Telnyx phone number (idempotent)",
        "security": [{ "GatewayToken": [] }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "phone": { "type": "string" }, "provisioned": { "type": "boolean" } } } } } } }
      }
    },
    "/api/proxy/email/send": {
      "post": {
        "tags": ["Proxy"],
        "summary": "Send an email via instance's inbox",
        "security": [{ "GatewayToken": [] }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "description": "AgentMail send message body" } } } },
        "responses": { "200": { "description": "Proxied from AgentMail" } }
      }
    },
    "/api/proxy/email/messages": {
      "get": {
        "tags": ["Proxy"],
        "summary": "Poll inbox messages",
        "security": [{ "GatewayToken": [] }],
        "responses": { "200": { "description": "Proxied from AgentMail" } }
      }
    },
    "/api/proxy/email/threads": {
      "get": {
        "tags": ["Proxy"],
        "summary": "Poll inbox threads",
        "security": [{ "GatewayToken": [] }],
        "responses": { "200": { "description": "Proxied from AgentMail" } }
      }
    },
    "/api/proxy/sms/send": {
      "post": {
        "tags": ["Proxy"],
        "summary": "Send SMS from instance's phone number",
        "security": [{ "GatewayToken": [] }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "to": { "type": "string" }, "text": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Proxied from Telnyx" } }
      }
    },
    "/api/proxy/sms/records": {
      "get": {
        "tags": ["Proxy"],
        "summary": "Poll inbound SMS records",
        "security": [{ "GatewayToken": [] }],
        "responses": { "200": { "description": "Proxied from Telnyx (filtered to instance's phone)" } }
      }
    },
    "/api/proxy/sms/messages/{id}": {
      "get": {
        "tags": ["Proxy"],
        "summary": "Get SMS message details",
        "security": [{ "GatewayToken": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Proxied from Telnyx" }, "403": { "description": "Message does not belong to this instance" } }
      }
    },
...
  }
}